update erro display
